+++ src/main/java/lotto/domain/Playslip.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class Playslip {
