+++ src/test/java/lotto/domain/PlayslipTest.java
@@ -1,31 +1,15 @@
-package lotto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.stream.Collectors;
+import lotto.Fixtures;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-public class PickedNumbersTest {
-
-    @Test
-    @DisplayName("6개의 로또 번호로 로또 번호 일급 컬렉션을 생성할 수 있다.")
-    void create() {
-        assertThat(Fixtures.createNumbers(Arrays.asList(1, 2, 3, 4, 5, 6)).size())
-            .isEqualTo(PickedNumbers.SIZE);
-    }
-
-    @Test
-    @DisplayName("로또 번호 갯수는 6보다 작거나 클 수 없다.")
-    void create_invalidSize() {
-        assertThatThrownBy(() -> new PickedNumbers(Collections.emptyList()))
-            .isInstanceOf(IllegalArgumentException.class);
-    }
+public class PlayslipTest {
 
     @ParameterizedTest
     @CsvSource(value = {
@@ -39,13 +23,14 @@ public class PickedNumbersTest {
     }, delimiter = ':')
     @DisplayName("로또 번호 갯수가 일치하는지 확인할 수 있다.")
     void contains(final int expected, final String numbers) {
-        final PickedNumbers pickedNumbers =
+        final Playslip playslip = new Playslip(
             Fixtures.createNumbers(
                 Arrays
                     .stream(numbers.split(","))
                     .map(Integer::parseInt)
                     .collect(Collectors.toList())
-            );
-        assertThat(pickedNumbers.contains(Fixtures.winningNumbers, expected)).isTrue();
+            )
+        );
+        assertThat(playslip.contains(Fixtures.winningNumbers, expected)).isTrue();
     }
 }
